+++ HISTORY.rst
@@ -18,6 +18,7 @@ New features and enhancements
   * A CLI version of this function is also available from the command line: `$ xclim show_version_info`. (:pull:`1073`).
 * New "keep_attrs" option to control the handling of the attributes within the indicators. (:issue:`1026`, :pull:`1076`).
 * Added a notebook showcasing some simple examples of Spatial Analogues. (:issue:`585`, :pull:`1075`).
+* ``create_ensembles`` now accepts a glob string to find datasets. (:pull:`1081`).
 
 Internal changes
 ^^^^^^^^^^^^^^^^